Update bot.py
added a few more specific response optimizations and their subsequent queries, such as QandA to elicit a response specifically catered to a question and answer scenario, currently only question as input. will be making a choice on Q or A prompt next reision. 

a data parsing centring prompt has been added, specify the objects via list input, describe some context, input the target string to be parsed. 

language translation queries, input text and request 1 or multiple languages for it to be translated to

@ query is now usable, i have made it so that the @ query specifically elicits a near completely randomized response from a randomly selected model with all variables randomized within range